# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axiom Lumen Public API",
    "version": "1.0.0",
    "description": "Persisted Stellar reconciliation snapshots. The production specification includes implemented routes only.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/api/v1/anchors/{anchor}/reserves": {
      "get": {
        "operationId": "getAnchorReserves",
        "tags": [
          "Anchors"
        ],
        "summary": "Get reviewed public reserve disclosures for a verified anchor",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Returns only publication-approved named-party flags and public corrections. An empty disclosures array does not reveal whether internal cases exist.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Anchor"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Public disclosures, possibly empty",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnchorReservesResponse"
                },
                "examples": {
                  "reviewedComparison": {
                    "$ref": "#/components/examples/AnchorReserves"
                  }
                }
              }
            }
          },
          "304": {
            "description": "The representation matches If-None-Match",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            }
          },
          "400": {
            "description": "Invalid anchor, pagination, request header, or query parameter",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "invalidAnchor": {
                    "$ref": "#/components/examples/InvalidAnchor"
                  },
                  "invalidPagination": {
                    "$ref": "#/components/examples/InvalidPagination"
                  },
                  "invalidRequestId": {
                    "$ref": "#/components/examples/InvalidRequestId"
                  },
                  "invalidQueryParameter": {
                    "$ref": "#/components/examples/InvalidQueryParameter"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationError"
          },
          "404": {
            "description": "No verified anchor exists for the identifier",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "missing": {
                    "$ref": "#/components/examples/AnchorMissing"
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "503": {
            "description": "The persisted public read model or API access storage is unavailable",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/XRateLimitReset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "readUnavailable": {
                    "$ref": "#/components/examples/AnchorReadUnavailable"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "operationId": "anchorReservesOptions",
        "summary": "Inspect public read-only CORS policy",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight accepted",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "Access-Control-Allow-Headers": {
                "$ref": "#/components/headers/CorsAllowHeaders"
              },
              "Access-Control-Allow-Methods": {
                "$ref": "#/components/headers/CorsAllowMethods"
              },
              "Access-Control-Max-Age": {
                "$ref": "#/components/headers/CorsMaxAge"
              }
            }
          },
          "400": {
            "description": "Invalid request identifier",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControl"
              },
              "Access-Control-Allow-Origin": {
                "$ref": "#/components/headers/CorsOrigin"
              },
              "Access-Control-Expose-Headers": {
                "$ref": "#/components/headers/CorsExposeHeaders"
              },
              "Vary": {
                "$ref": "#/components/headers/Vary"
              },
              "Access-Control-Allow-Headers": {
                "$ref": "#/components/headers/CorsAllowHeaders"
              },
              "Access-Control-Allow-Methods": {
                "$ref": "#/components/headers/CorsAllowMethods"
              },
              "Access-Control-Max-Age": {
                "$ref": "#/components/headers/CorsMaxAge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "invalidRequestId": {
                    "$ref": "#/components/examples/InvalidRequestId"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]*$"
          },
          "as_of": {
            "type": "string",
            "format": "date-time"
          },
          "api_version": {
            "type": "string",
            "const": "v1"
          }
        },
        "required": [
          "error",
          "request_id",
          "as_of",
          "api_version"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
      "AuthenticationError": {
        "description": "API key is missing, malformed, expired, revoked, or inactive when authentication is required",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControl"
          },
          "Access-Control-Allow-Origin": {
            "$ref": "#/components/headers/CorsOrigin"
          },
          "Access-Control-Expose-Headers": {
            "$ref": "#/components/headers/CorsExposeHeaders"
          },
          "Vary": {
            "$ref": "#/components/headers/Vary"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "examples": {
              "authentication": {
                "$ref": "#/components/examples/AuthenticationError"
              }
            }
          }
        }
      },
      "RateLimitError": {
        "description": "The authenticated principal exhausted its plan quota for the current fixed window",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControl"
          },
          "Access-Control-Allow-Origin": {
            "$ref": "#/components/headers/CorsOrigin"
          },
          "Access-Control-Expose-Headers": {
            "$ref": "#/components/headers/CorsExposeHeaders"
          },
          "Vary": {
            "$ref": "#/components/headers/Vary"
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          },
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/XRateLimitLimit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/XRateLimitRemaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/XRateLimitReset"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "examples": {
              "rateLimit": {
                "$ref": "#/components/examples/RateLimitError"
              }
            }
          }
        }
      }
    },
    "parameters": {
...
      "Anchor": {
        "name": "anchor",
        "in": "path",
        "required": true,
        "description": "Canonical verified anchor identifier",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]*$"
        },
        "example": "anchor_example"
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the preceding anchor disclosure page",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Disclosure page size; defaults to 25 and cannot exceed 100",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional caller correlation identifier; generated when absent",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]*$"
        }
      },
      "IfNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Request correlation identifier",
        "schema": {
          "type": "string"
        }
      },
      "CacheControl": {
        "description": "Private snapshot caching or no-store",
        "schema": {
          "type": "string"
        }
      },
      "ETag": {
        "description": "Weak complete-representation validator",
        "schema": {
          "type": "string"
        }
      },
      "CorsOrigin": {
        "description": "Public read-only origin policy",
        "schema": {
          "type": "string",
          "enum": [
            "*"
          ]
        }
      },
      "CorsExposeHeaders": {
        "description": "Response headers available to browsers",
        "schema": {
          "type": "string"
        }
      },
      "CorsAllowHeaders": {
        "description": "Headers accepted by CORS requests",
        "schema": {
          "type": "string"
        }
      },
      "CorsAllowMethods": {
        "description": "Methods accepted by CORS requests",
        "schema": {
          "type": "string",
          "enum": [
            "GET, OPTIONS"
          ]
        }
      },
      "CorsMaxAge": {
        "description": "Preflight cache lifetime in seconds",
        "schema": {
          "type": "string",
          "enum": [
            "86400"
          ]
        }
      },
      "Vary": {
        "description": "Cache representation varies by request identifier and API credential",
        "schema": {
          "type": "string",
          "enum": [
            "X-Request-ID, X-Axiom-Key"
          ]
        }
      },
      "RetryAfter": {
        "description": "Seconds until a quota-limited request may be retried",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
      "XRateLimitLimit": {
        "description": "Maximum requests in the current plan window",
        "schema": {
          "type": "string",
          "pattern": "^[1-9]\\d*$"
        }
      },
      "XRateLimitRemaining": {
        "description": "Requests remaining after this request",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      },
      "XRateLimitReset": {
        "description": "UTC Unix timestamp when the fixed window resets",
        "schema": {
          "type": "string",
          "pattern": "^\\d+$"
        }
      }
    },
    "examples": {
...
      "AnchorReserves": {
        "summary": "Reviewed reserve comparison disclosure",
        "value": {
          "anchor": {
            "id": "anchor_example",
            "name": "Example Anchor",
            "network": "public",
            "stellar_account": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
            "status": "verified"
          },
          "disclosures": [
            {
              "flag_id": "anchor_flag_example",
              "severity": "warning",
              "lifecycle_state": "open",
              "publication_state": "approved_public",
              "methodology_version": "anchor-reserve-comparison-v0.1",
              "approved_at": "2026-08-10T12:00:00.000Z",
              "first_observed_at": "2026-08-10T12:00:00.000Z",
              "last_observed_at": "2026-08-10T12:00:00.000Z",
              "measurement": {
                "event_id": "anchor_measurement_example",
                "measured_at": "2026-08-10T12:00:00.000Z",
                "asset": "USDC:GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
                "reserve_amount": "970",
                "onchain_supply": "1000",
                "absolute_delta": "30",
                "delta_basis_points": 300,
                "attestation_period_start": "2026-08-10T11:00:00.000Z",
                "attestation_period_end": "2026-08-10T12:00:00.000Z",
                "published_at": "2026-08-10T12:00:00.000Z",
                "attestation": {
                  "schema": "axiom-lumen-anchor-reserve-attestation-v1",
                  "document_url": "https://anchor.example/reserves",
                  "evidence_sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                },
                "source": {
                  "id": "anchor_source_example",
                  "url": "https://anchor.example/reserves",
                  "source_class": "anchor_self_reported"
                },
                "supply_reference": {
                  "snapshot_id": "supply_snapshot_example",
                  "amount": "1000",
                  "as_of": "2026-08-10T12:00:00.000Z",
                  "ledger_sequence": 58000000,
                  "ledger_closed_at": "2026-08-10T12:00:00.000Z",
                  "status": "verified",
                  "confidence": 0.95,
                  "methodology_version": "onchain-asset-supply-v0.1"
                },
                "confidence": {
                  "score": 0.49,
                  "formula_version": "anchor-reserve-confidence-v0.1",
                  "components": {
                    "attestation": 1,
                    "reference": 0.95,
                    "temporal_alignment": 1
                  },
                  "caps_applied": [
                    "anchor_self_reported"
                  ]
                }
              },
              "response": null,
              "disputes": [],
              "corrections": []
            }
          ],
          "page": {
            "next_cursor": null
          },
          "as_of": "2026-08-10T12:00:00.000Z",
          "request_id": "example_request",
          "api_version": "v1"
        }
      },
      "InvalidRequestId": {
        "summary": "Invalid request identifier",
        "value": {
          "error": {
            "code": "invalid_request_id",
            "message": "X-Request-ID must be a valid identifier of at most 128 characters"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "InvalidQueryParameter": {
        "summary": "Unsupported query parameter",
        "value": {
          "error": {
            "code": "invalid_query_parameter",
            "message": "Unsupported query parameter: limit"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "InvalidAnchor": {
        "summary": "Invalid anchor identifier",
        "value": {
          "error": {
            "code": "invalid_anchor",
            "message": "Anchor must be a valid canonical identifier"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "InvalidPagination": {
        "summary": "Invalid anchor disclosure pagination",
        "value": {
          "error": {
            "code": "invalid_pagination",
            "message": "anchor reserve cursor is invalid"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "AnchorMissing": {
        "summary": "No verified anchor",
        "value": {
          "error": {
            "code": "anchor_not_found",
            "message": "No verified anchor is available for this identifier"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
...
      "AnchorReadUnavailable": {
        "summary": "Anchor reserve read storage unavailable",
        "value": {
          "error": {
            "code": "anchor_reserves_read_unavailable",
            "message": "The anchor reserve read model is temporarily unavailable"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "AuthenticationError": {
        "summary": "Authentication is required but the supplied API key is not usable",
        "value": {
          "error": {
            "code": "authentication_required",
            "message": "A valid API key is required"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      },
      "RateLimitError": {
        "summary": "The current API-key plan window is exhausted",
        "value": {
          "error": {
            "code": "rate_limit_exceeded",
            "message": "The request quota has been exceeded"
          },
          "request_id": "example_request",
          "as_of": "2026-08-10T12:00:00.000Z",
          "api_version": "v1"
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Axiom-Key",
        "description": "Required by the hosted API contract. Plaintext keys are shown only when issued."
      }
    }
  }
}